requirements.txt
flask
Flask-RESTful
gunicorn


cloudshell-automation-api
